Add: Files for each class, Fix: draw and winning validation
