+++ lib/player.rb
@@ -1,0 +1,15 @@
+# Create players
+
+class Player
+  attr_reader :name, :mark, :pick, :choices
+
+  @@count = 0
+
+  def initialize(name)
+    @name = name
+    @mark = @@count.zero? ? '+' : '-'
+    @pick = pick
+    @choices = []
+    @@count += 1
+  end
+end
